wip: add more signals to app context
