+++ src/contexts/app.context.tsx
@@ -3,14 +3,31 @@ import {
     JSXElement,
     Setter,
     createContext,
+    createEffect,
     createSignal,
     useContext,
 } from "solid-js";
+import { OccurrenceFiles } from "../models";
+import {
+    setIsLoadingSelectedOccurrence,
+    setIsPlaying,
+    setSelectedOccurrence,
+} from "../stores";
 
 export interface AppContextInterface {
     platform: {
         isDesktop: Accessor<boolean>;
         setIsDesktop: Setter<boolean>;
+    };
+    sidebar: {
+        isOpen: Accessor<boolean>;
+        setIsOpen: Setter<boolean>;
+    };
+    fileByOccurrence: {
+        get: Accessor<OccurrenceFiles[]>;
+        set: Setter<OccurrenceFiles[]>;
+        getSelected: Accessor<OccurrenceFiles | null>;
+        setSelected: Setter<OccurrenceFiles | null>;
     };
 }
 
@@ -22,11 +39,63 @@ export interface AppProviderInterface {
 
 export const AppProvider = (props: AppProviderInterface) => {
     const [isDesktop, setIsDesktop] = createSignal(false);
+    const [isSidebarOpen, setIsSidebarOpen] = createSignal(false);
+
+    const [fileByOccurrence, setFilesByOccurrences] = createSignal<
+        Array<OccurrenceFiles>
+    >([]);
+    const [selectedOccurrenceFiles, setSelectedOccurrenceFiles] =
+        createSignal<OccurrenceFiles | null>(null);
+
+    createEffect(() => {
+        const _selectedOccurrenceFiles = selectedOccurrenceFiles();
+
+        if (!_selectedOccurrenceFiles) {
+            return;
+        }
+
+        setSelectedOccurrence(null);
+        setIsLoadingSelectedOccurrence(true);
+        setIsPlaying(false);
+        // TODO: close
+        // setIsSidebarOpen(false);
+
+        _selectedOccurrenceFiles
+            .toOccurrence()
+            .then((occurrence) => {
+                if (!occurrence) {
+                    console.error(
+                        "Failed to convert OccurenceFiles to Occurence"
+                    );
+                    return;
+                }
+
+                setIsLoadingSelectedOccurrence(false);
+                setSelectedOccurrence(occurrence);
+            })
+            .catch((error) => {
+                setIsLoadingSelectedOccurrence(false);
+                console.error(
+                    "Error converting occurrenceFiles to Occurrence",
+                    error
+                );
+            });
+    });
 
     const app: AppContextInterface = {
         platform: {
             isDesktop,
             setIsDesktop,
+        },
+        sidebar: {
+            isOpen: isSidebarOpen,
+            setIsOpen: setIsSidebarOpen,
+        },
+        fileByOccurrence: {
+            get: fileByOccurrence,
+            set: setFilesByOccurrences,
+            getSelected: selectedOccurrenceFiles,
+            setSelected: setSelectedOccurrenceFiles,
         },
     };
