+++ src/models/index.ts
@@ -1,4 +1,5 @@
 export * from './occurence';
+export * from './occurence-files';
 export * from './config';
 export * from './camera-position';
 export * from './player-start-point';
