--- src/stores/footer.store.ts
@@ -1,3 +1,0 @@
-import { createSignal } from "solid-js";
-
-export const [isLoadingOccurrences, setIsLoadingOccurrences] = createSignal(false);
