+++ src/stores/index.tsx
@@ -1,3 +1,2 @@
-export * from "./footer.store";
 export * from "./occurrences.store";
 export * from "./sidebar.store";
