+++ src/stores/sidebar.store.ts
@@ -1,7 +1,5 @@
 import { createSignal } from "solid-js";
 
-export const [isSidebarOpen, setIsSidebarOpen] = createSignal(false);
-
 export enum Filter {
     All = "all",
     Sentry = "sentry",
