+++ src/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './to-base64';
 export * from './tauri';
 export * from './uuid';
-export * from './add-folder';
